Feat: 별점 기능 추가

+++ src/main/java/com/aws/compass/entity/ReviewCount.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Data
 public class ReviewCount {
     private int reviewCount;
-    private int scoreAvg;
+    private float scoreAvg;
 }
